Inline mapping tests into main RrepwMappersTests
